refactor(downgrade): change (un)marshaling to copy
